+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## git version
 
+- Deprecated `Deferred.choice` type alias.  Use `Deferred.Choice.t`
+  instead.
+
 - Removed deprecated `ignore` functions from `Deferred`,
   `Deferred.Result`, and `Deferred.Or_error` (and the corresponding
   `Eager_deferred` modules).